moved shared methods between install_activity and github_activity to utils.py and modified install_activity_model.py and github_activity_model.py to reflect the changes
